pyproject.toml
[tool.poetry]
name = "django-rest-framwork"
version = "0.1.0"
description = ""
authors = ["waelelsafty07 <waelelsafty07@gmail.com>"]
readme = "README.md"
packages = [{include = "django_rest_framwork"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
yapf = "^0.32.0"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
